+++ crates/rust-client/src/note/note_screener.rs
@@ -1,22 +1,29 @@
-use alloc::{collections::BTreeSet, vec::Vec};
+use alloc::{sync::Arc, vec::Vec};
 use core::fmt;
 
-use miden_lib::note::well_known_note::WellKnownNote;
+use miden_lib::{account::interface::AccountInterface, note::well_known_note::WellKnownNote};
 use miden_objects::{
-    AccountError, AssetError, Felt, Word,
+    AccountError, AssetError,
     account::{Account, AccountId},
-    asset::Asset,
     note::{Note, NoteId},
+    transaction::{InputNote, InputNotes},
+};
+use miden_tx::{
+    NoteAccountExecution, NoteConsumptionChecker, TransactionExecutor, TransactionExecutorError,
+    TransactionMastStore,
 };
 use thiserror::Error;
 
-use crate::store::{Store, StoreError};
+use crate::{
+    store::{Store, StoreError},
+    transaction::{TransactionRequestBuilder, TransactionRequestError},
+};
 
 /// Describes the relevance of a note based on the screening.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub enum NoteRelevance {
-    /// The note can be consumed at any time.
-    Always,
+    /// The note can be consumed in the current block.
+    Now,
     /// The note can be consumed after the block with the specified number.
     After(u32),
 }
@@ -30,7 +37,7 @@ pub type NoteConsumability = (AccountId, NoteRelevance);
 impl fmt::Display for NoteRelevance {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            NoteRelevance::Always => write!(f, "Always"),
+            NoteRelevance::Now => write!(f, "Now"),
             NoteRelevance::After(height) => write!(f, "After block {height}"),
         }
     }
@@ -42,13 +49,26 @@ impl fmt::Display for NoteRelevance {
 /// tracked in the provided `store`. This can be derived in a number of ways, such as looking
 /// at the combination of script root and note inputs. For example, a P2ID note is relevant
 /// for a specific account ID if this ID is its first note input.
-pub struct NoteScreener {
-    store: alloc::sync::Arc<dyn Store>,
+pub struct NoteScreener<'a> {
+    /// A reference to the client's store, used to fetch necessary data to check consumability.
+    store: Arc<dyn Store>,
+    /// A consumability checker, used to check whether a note can be consumed by an account.
+    consumability_checker: NoteConsumptionChecker<'a>,
+    /// A MAST store, used to provide code inputs to the VM.
+    mast_store: Arc<TransactionMastStore>,
 }
 
-impl NoteScreener {
-    pub fn new(store: alloc::sync::Arc<dyn Store>) -> Self {
-        Self { store }
+impl<'a> NoteScreener<'a> {
+    pub fn new(
+        store: Arc<dyn Store>,
+        tx_executor: &'a TransactionExecutor,
+        mast_store: Arc<TransactionMastStore>,
+    ) -> Self {
+        Self {
+            store,
+            consumability_checker: NoteConsumptionChecker::new(tx_executor),
+            mast_store,
+        }
     }
 
     /// Returns a vector of tuples describing the relevance of the provided note to the
@@ -61,65 +81,78 @@ impl NoteScreener {
         &self,
         note: &Note,
     ) -> Result<Vec<NoteConsumability>, NoteScreenerError> {
-        let account_ids = BTreeSet::from_iter(self.store.get_account_ids().await?);
+        let mut note_relevances = vec![];
+        for id in self.store.get_account_ids().await? {
+            let account_record = self
+                .store
+                .get_account(id)
+                .await?
+                .ok_or(NoteScreenerError::AccountDataNotFound(id))?;
 
-        let script_root = note.script().root();
+            if let Some(relevance) =
+                self.check_standard_consumability(account_record.account(), note).await?
+            {
+                note_relevances.push((id, relevance));
+            } else {
+                // The note might be consumable after a certain block height if the note is p2idr
+                let script_root = note.script().root();
 
-        let note_relevance = match script_root {
-            ref p2id_root if p2id_root == &WellKnownNote::P2ID.script_root() => {
-                Self::check_p2id_relevance(note, &account_ids)?
-            },
-            ref p2idr_root if p2idr_root == &WellKnownNote::P2IDR.script_root() => {
-                Self::check_p2idr_relevance(note, &account_ids)?
-            },
-            ref swap_root if swap_root == &WellKnownNote::SWAP.script_root() => {
-                self.check_swap_relevance(note, &account_ids).await?
-            },
-            _ => NoteScreener::check_script_relevance(note, &account_ids),
-        };
-
-        Ok(note_relevance)
-    }
-
-    fn check_p2id_relevance(
-        note: &Note,
-        account_ids: &BTreeSet<AccountId>,
-    ) -> Result<Vec<NoteConsumability>, NoteScreenerError> {
-        let note_inputs = note.inputs().values();
-        if note_inputs.len() != 2 {
-            return Err(InvalidNoteInputsError::WrongNumInputs(note.id(), 2).into());
+                if script_root == WellKnownNote::P2IDR.script_root() {
+                    if let Some(relevance) = Self::check_p2idr_recall_consumability(note, &id)? {
+                        note_relevances.push((id, relevance));
+                    }
+                }
+            }
         }
 
-        let account_id_felts: [Felt; 2] = note_inputs[0..2].try_into().expect(
-            "Should be able to convert the first two note inputs to an array of two Felt elements",
-        );
-
-        let account_id =
-            AccountId::try_from([account_id_felts[1], account_id_felts[0]]).map_err(|err| {
-                InvalidNoteInputsError::AccountError(
-                    note.id(),
-                    AccountError::FinalAccountHeaderIdParsingFailed(err),
-                )
-            })?;
-
-        if !account_ids.contains(&account_id) {
-            return Ok(vec![]);
-        }
-        Ok(vec![(account_id, NoteRelevance::Always)])
+        Ok(note_relevances)
     }
 
-    fn check_p2idr_relevance(
+    /// Tries to execute a standard consume transaction to check if the note is consumable by the
+    /// account.
+    async fn check_standard_consumability(
+        &self,
+        account: &Account,
         note: &Note,
-        account_ids: &BTreeSet<AccountId>,
-    ) -> Result<Vec<NoteConsumability>, NoteScreenerError> {
+    ) -> Result<Option<NoteRelevance>, NoteScreenerError> {
+        let transaction_request =
+            TransactionRequestBuilder::consume_notes(vec![note.id()]).build()?;
+
+        let tx_script =
+            transaction_request.build_transaction_script(&AccountInterface::from(account), true)?;
+
+        let tx_args = transaction_request.clone().into_transaction_args(tx_script, vec![]);
+        let input_notes = InputNotes::new(vec![InputNote::unauthenticated(note.clone())])
+            .expect("Single note should be valid");
+
+        self.mast_store.load_transaction_code(account.code(), &input_notes, &tx_args);
+
+        if let NoteAccountExecution::Success = self
+            .consumability_checker
+            .check_notes_consumability(
+                account.id(),
+                self.store.get_sync_height().await?,
+                input_notes,
+                tx_args,
+            )
+            .await?
+        {
+            return Ok(Some(NoteRelevance::Now));
+        }
+
+        Ok(None)
+    }
+
+    /// Special relevance check for P2IDR notes. It checks if the sender account can consume and
+    /// recall the note.
+    fn check_p2idr_recall_consumability(
+        note: &Note,
+        account_id: &AccountId,
+    ) -> Result<Option<NoteRelevance>, NoteScreenerError> {
         let note_inputs = note.inputs().values();
         if note_inputs.len() != 3 {
             return Err(InvalidNoteInputsError::WrongNumInputs(note.id(), 3).into());
         }
-
-        let account_id_felts: [Felt; 2] = note_inputs[0..2].try_into().expect(
-            "Should be able to convert the first two note inputs to an array of two Felt elements",
-        );
 
         let recall_height_felt = note_inputs[2];
 
@@ -128,97 +161,11 @@ impl NoteScreener {
             InvalidNoteInputsError::BlockNumberError(note.id(), recall_height_felt.as_int())
         })?;
 
-        let account_id =
-            AccountId::try_from([account_id_felts[1], account_id_felts[0]]).map_err(|err| {
-                InvalidNoteInputsError::AccountError(
-                    note.id(),
-                    AccountError::FinalAccountHeaderIdParsingFailed(err),
-                )
-            })?;
-
-        Ok(vec![
-            (account_id, NoteRelevance::Always),
-            (sender, NoteRelevance::After(recall_height)),
-        ]
-        .into_iter()
-        .filter(|(account_id, _relevance)| account_ids.contains(account_id))
-        .collect())
-    }
-
-    /// Checks if a swap note can be consumed by any account whose ID is in `account_ids`.
-    ///
-    /// This implementation serves as a placeholder as we're currently not able to create, execute
-    /// and send SWAP NOTES. Hence, it's also untested. The main logic should be the same: for each
-    /// account check if it has enough of the wanted asset.
-    /// This is also very inefficient as we're loading the full accounts. We should instead just
-    /// load the account's vaults, or even have a function in the `Store` to do this.
-    // TODO: test/revisit this in the future
-    async fn check_swap_relevance(
-        &self,
-        note: &Note,
-        account_ids: &BTreeSet<AccountId>,
-    ) -> Result<Vec<NoteConsumability>, NoteScreenerError> {
-        let note_inputs = note.inputs().values();
-        if note_inputs.len() != 10 {
-            return Err(InvalidNoteInputsError::WrongNumInputs(note.id(), 10).into());
+        if sender == *account_id {
+            Ok(Some(NoteRelevance::After(recall_height)))
+        } else {
+            Ok(None)
         }
-
-        let asset_felts: [Felt; 4] = note_inputs[4..8].try_into().expect(
-            "Should be able to convert the second word from note inputs to an array of four Felt elements",
-        );
-
-        // get the demanded asset from the note's inputs
-        let asset: Asset = Word::from(asset_felts)
-            .try_into()
-            .map_err(|err| InvalidNoteInputsError::AssetError(note.id(), err))?;
-
-        let mut accounts_with_relevance = Vec::new();
-
-        for account_id in account_ids {
-            let account: Account = self
-                .store
-                .get_account(*account_id)
-                .await?
-                .ok_or(NoteScreenerError::AccountDataNotFound(*account_id))?
-                .into();
-
-            // Check that the account can cover the demanded asset
-            match asset {
-                Asset::NonFungible(non_fungible_asset)
-                    if account.vault().has_non_fungible_asset(non_fungible_asset).expect(
-                        "Should be able to query has_non_fungible_asset for an Asset::NonFungible",
-                    ) =>
-                {
-                    accounts_with_relevance.push((*account_id, NoteRelevance::Always));
-                },
-                Asset::Fungible(fungible_asset) => {
-                    let asset_faucet_id = fungible_asset.faucet_id();
-                    if account
-                        .vault()
-                        .get_balance(asset_faucet_id)
-                        .expect("Should be able to query get_balance for an Asset::Fungible")
-                        >= fungible_asset.amount()
-                    {
-                        accounts_with_relevance.push((*account_id, NoteRelevance::Always));
-                    }
-                },
-                Asset::NonFungible(_) => {},
-            }
-        }
-
-        Ok(accounts_with_relevance)
-    }
-
-    fn check_script_relevance(
-        _note: &Note,
-        account_ids: &BTreeSet<AccountId>,
-    ) -> Vec<NoteConsumability> {
-        // TODO: try to execute the note script against relevant accounts; this will
-        // require querying data from the store
-        account_ids
-            .iter()
-            .map(|account_id| (*account_id, NoteRelevance::Always))
-            .collect()
     }
 }
 
@@ -234,6 +181,10 @@ pub enum NoteScreenerError {
     AccountDataNotFound(AccountId),
     #[error("error while fetching data from the store")]
     StoreError(#[from] StoreError),
+    #[error("error while checking consume transaction")]
+    TransactionExecutionError(#[from] TransactionExecutorError),
+    #[error("error while building consume transaction request")]
+    TransactionRequestError(#[from] TransactionRequestError),
 }
 
 #[derive(Debug, Error)]