+++ crates/web-client/src/sync.rs
@@ -17,4 +17,18 @@ impl WebClient {
             Err(JsValue::from_str("Client not initialized"))
         }
     }
+
+    #[wasm_bindgen(js_name = "getSyncHeight")]
+    pub async fn get_sync_height(&mut self) -> Result<u32, JsValue> {
+        if let Some(client) = self.get_mut_inner() {
+            let sync_height = client
+                .get_sync_height()
+                .await
+                .map_err(|err| js_error_with_context(err, "failed to get sync height"))?;
+
+            Ok(sync_height.as_u32())
+        } else {
+            Err(JsValue::from_str("Client not initialized"))
+        }
+    }
 }